requirements.txt
numpy
fastapi
pydantic
uvicorn
torch
torchvision
torchaudio
Pillow

